fix: cast statvfs fields to u64 for macOS compatibility
On macOS, libc::statvfs f_bavail is u32 and f_bsize is u64.
On Linux both are u64. Cast both to u64 explicitly to compile
on all platforms.

+++ layers/compute/src/disk.rs
@@ -306,7 +306,7 @@ fn available_disk_space(path: &Path) -> u64 {
     unsafe {
         let mut stat: libc::statvfs = std::mem::zeroed();
         if libc::statvfs(c_path.as_ptr(), &mut stat) == 0 {
-            stat.f_bavail * stat.f_bsize
+            (stat.f_bavail as u64) * (stat.f_bsize as u64)
         } else {
             u64::MAX // cannot check — assume enough
         }
